% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modl.R
\name{modl.arima}
\alias{modl.arima}
\title{Automatic ARIMA model}
\usage{
modl.arima(tserie)
}
\arguments{
\item{tserie}{A ts object.}
}
\value{
ARIMA model.
}
\description{
Assuming "tserie" is stationary, returns the best arima model
}
\examples{
modl.arima(AirPassengers)
}
\author{
Alberto Vico Moreno
}
