% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modl.R
\name{modl.tsToDataFrame}
\alias{modl.tsToDataFrame}
\title{Ts to data frame transformation}
\usage{
modl.tsToDataFrame(tserie, formula = NULL)
}
\arguments{
\item{tserie}{A ts object.}

\item{formula}{An integer vector. Contains the indexes from the \code{tserie} wich will indicate how to extract the features. The last value will be the class index. Default value: c(1:16). Has to be length 6 minimum.}
}
\value{
the time serie as data frame
}
\description{
Transform a ts object into a data frame using the given formula.
}
\examples{
modl.tsToDataFrame(AirPassengers,formula=c(1,3,4,5,6,7))
modl.tsToDataFrame(AirPassengers,formula=c(1:20))
}
\author{
Alberto Vico Moreno
}
