% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred.R
\name{plot.pred}
\alias{plot.pred}
\title{Generic function}
\usage{
\method{plot}{pred}(x, ylab = "Values", main = "Predictions", ...)
}
\arguments{
\item{x}{\code{pred} object}

\item{ylab}{ylab}

\item{main}{main}

\item{...}{ignored}
}
\description{
Plots object prep
}
\examples{
plot(pred(modl(prep(AirPassengers))))
}
