% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postp.R
\name{postp.detrend.differencing}
\alias{postp.detrend.differencing}
\title{Undo detrend(differencing)}
\usage{
postp.detrend.differencing(tserie, nd, firstvalues)
}
\arguments{
\item{tserie}{A \code{ts} object.}

\item{nd}{Number of differences.}

\item{firsvalues}{Values lost on the original differences}
}
\value{
A \code{ts} object.
}
\description{
Uses inverse differences to revert the changes
}
\author{
Alberto Vico Moreno
}
