% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postp.R
\name{postp.detrend.sfsm}
\alias{postp.detrend.sfsm}
\title{Undo detrend(substracting full-means method)}
\usage{
postp.detrend.sfsm(tserie, means, start, frequency)
}
\arguments{
\item{tserie}{A \code{ts} object.}

\item{means}{A numeric vector.}

\item{start}{Start of original time serie}

\item{frequency}{Frequency of the original time serie}
}
\value{
A \code{ts} object.
}
\description{
Undo detrend(substracting full-means method)
}
\examples{
p <- prep.detrend.sfsm(AirPassengers)
postp.detrend.sfsm(p$tserie,p$means,start(AirPassengers),frequency(AirPassengers))
}
\author{
Alberto Vico Moreno
}
