% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postp.R
\name{postp.homogenize.boxcox}
\alias{postp.homogenize.boxcox}
\title{Undo Box-Cox transformation}
\usage{
postp.homogenize.boxcox(tserie, lambda)
}
\arguments{
\item{tserie}{A \code{ts} object.}

\item{lambda}{A numeric.}
}
\value{
A \code{ts} object.
}
\description{
Undo Box-Cox transformation
}
\examples{
p <- prep.homogenize.boxcox(AirPassengers)
postp.homogenize.boxcox(p$tserie,p$lambda)
}
\author{
Alberto Vico Moreno
}
