% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postp.R
\name{postp.homogenize.log}
\alias{postp.homogenize.log}
\title{Undo logaritmic transformation}
\usage{
postp.homogenize.log(tserie)
}
\arguments{
\item{tserie}{A \code{ts} object.}
}
\value{
A \code{ts} object.
}
\description{
Uses exponent to reverse the logarithm
}
\examples{
postp.homogenize.log(prep.homogenize.log(AirPassengers))
}
\author{
Alberto Vico Moreno
}
