% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred.R
\name{pred.arima}
\alias{pred.arima}
\title{Predicts for ARIMA}
\usage{
pred.arima(model, n.ahead)
}
\arguments{
\item{model}{An ARIMA model.}

\item{n.ahead}{Number of values to predict.}
}
\value{
A \code{ts} object containing the predictions.
}
\description{
Performs predictions over an ARIMA model using the \code{stats::predict} function.
}
\examples{
pred.arima(forecast::auto.arima(prep(AirPassengers)$tserie),n.ahead=30)
}
\author{
Alberto Vico Moreno
}
