% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred.R
\name{pred.dataMining}
\alias{pred.dataMining}
\title{Predicts for data mining methods}
\usage{
pred.dataMining(model, n.ahead)
}
\arguments{
\item{model}{A \code{modl} object.}

\item{n.ahead}{Number of values to predict.}
}
\value{
A \code{ts} object containing the predictions.
}
\description{
Performs predictions over a data mining model using the \code{caret::predict.train} function.
}
\examples{
m <- modl(prep(AirPassengers),method='dataMining',algorithm='rpart')
pred.dataMining(m,n.ahead=15)
}
\author{
Alberto Vico Moreno
}
