% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{prep.check.acf}
\alias{prep.check.acf}
\title{Autocorrelation function}
\usage{
prep.check.acf(tserie)
}
\arguments{
\item{tserie}{a \code{ts} or a \code{prep} object}
}
\description{
Plots the autocorrelation function to check stationarity
}
\details{
For a stationary time series, the ACF will drop to zero
relatively quickly, while the ACF of non-stationary data decreases slowly.
Also, for non-stationary data, the value is often large and positive.
}
\examples{
prep.check.acf(AirPassengers)
prep.check.acf(prep(AirPassengers))
}
