% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{prep.homogenize.boxcox}
\alias{prep.homogenize.boxcox}
\title{Box-Cox transformation}
\usage{
prep.homogenize.boxcox(tserie)
}
\arguments{
\item{tserie}{a \code{ts} object}
}
\value{
A list is returned containing:
\item{boxcox}{Transformed ts object.}
\item{lambda}{Lambda value.}
}
\description{
Performs a Box-Cox transformation to a time serie.
}
\examples{
prep.homogenize.log(AirPassengers)
}
\references{
Box-Cox transformation: \url{https://en.wikipedia.org/wiki/Power_transform#Box.E2.80.93Cox_transformation}
}
