% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{prep.homogenize.log}
\alias{prep.homogenize.log}
\title{Logaritmic transformation}
\usage{
prep.homogenize.log(tserie)
}
\arguments{
\item{tserie}{a \code{ts} object}
}
\value{
\code{ts} object with transformed time serie
}
\description{
Performs a logaritmic transformation to a time serie.
}
\examples{
prep.homogenize.log(AirPassengers)
}
