% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modl.R
\name{print.modl}
\alias{print.modl}
\title{Generic function}
\usage{
\method{print}{modl}(x, ...)
}
\arguments{
\item{x}{\code{prep} object}

\item{...}{ignored}
}
\description{
Prints object modl
}
\examples{
print(modl(prep(AirPassengers)))
}
