% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred.R
\name{print.pred}
\alias{print.pred}
\title{Generic function}
\usage{
\method{print}{pred}(x, ...)
}
\arguments{
\item{x}{\code{prep} object}

\item{...}{ignored}
}
\description{
Prints object pred
}
\examples{
print(pred(modl(prep(AirPassengers))))
}
