% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{print.prep}
\alias{print.prep}
\title{Generic function}
\usage{
\method{print}{prep}(x, ...)
}
\arguments{
\item{x}{\code{prep} object}

\item{...}{ignored}
}
\description{
Prints object prep
}
\examples{
print(prep(AirPassengers))
}
