% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modl.R
\name{summary.modl}
\alias{summary.modl}
\title{Generic function}
\usage{
\method{summary}{modl}(object, ...)
}
\arguments{
\item{object}{\code{prep} object}

\item{...}{ignored}
}
\description{
Summary of object modl
}
\examples{
summary(modl(prep(AirPassengers)))
}
