% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred.R
\name{summary.pred}
\alias{summary.pred}
\title{Generic function}
\usage{
\method{summary}{pred}(object, ...)
}
\arguments{
\item{object}{\code{prep} object}

\item{...}{ignored}
}
\description{
Summary of object pred
}
\examples{
summary(pred(modl(prep(AirPassengers))))
}
