% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{summary.prep}
\alias{summary.prep}
\title{Generic function}
\usage{
\method{summary}{prep}(object, ...)
}
\arguments{
\item{object}{\code{prep} object}

\item{...}{ignored}
}
\description{
Summary of object prep
}
\examples{
summary(prep(AirPassengers))
}
